Refactoring code. Fixing examples and License.
